pyproject.toml
[project]
name = "sales-leads-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.47",
    "llama-index-llms-google-genai>=0.2.3",
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
